Move app inside src directory
