+++ src/app/App.tsx
@@ -1,4 +1,5 @@
 function App() {
+  console.log('WELCOME TO CLAUDE');
   return (
     <div>
       <h1>Welcome to Claude Minions</h1>
