+++ src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import App from './App'
+import App from './app/App'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
