+++ src/vite.config.ts
